refactor: de-complexify the src/processors
