+++ deno.lock
@@ -680,9 +680,11 @@
     "https://deno.land/std@0.224.0/testing/mock.ts": "a963181c2860b6ba3eb60e08b62c164d33cf5da7cd445893499b2efda20074db",
     "https://deno.land/x/hono@v3.11.11/client/client.ts": "ff340f58041203879972dd368b011ed130c66914f789826610869a90603406bf",
     "https://deno.land/x/hono@v3.11.11/client/index.ts": "3ff4cf246f3543f827a85a2c84d66a025ac350ee927613629bda47e854bfb7ba",
+    "https://deno.land/x/hono@v3.11.11/client/types.ts": "a947b8c7ed25169062c4c93f2fd2042da617578aa0d13da05a1e8b8877e47f0f",
     "https://deno.land/x/hono@v3.11.11/client/utils.ts": "053273c002963b549d38268a1b423ac8ca211a8028bdab1ed0b781a62aa5e661",
     "https://deno.land/x/hono@v3.11.11/compose.ts": "37d6e33b7db80e4c43a0629b12fd3a1e3406e7d9e62a4bfad4b30426ea7ae4f1",
     "https://deno.land/x/hono@v3.11.11/context.ts": "5cac561498e1f4af7a355673a047f07f47349056c5167ed9aec8a20cdcac8464",
+    "https://deno.land/x/hono@v3.11.11/helper/adapter/index.ts": "eea9b4caedbfa3a3b4a020bf46c88c0171a00008cd6c10708cd85a3e39d86e62",
     "https://deno.land/x/hono@v3.11.11/helper/cookie/index.ts": "a05ce7e3bafe1f8c7f45d04abf79822716011be75904fe1aad2e99126fd985b9",
     "https://deno.land/x/hono@v3.11.11/helper/html/index.ts": "701ed12b808e8c253247edc02c2c4b32261ae899e9c98f1427f7f6eaa8b7d1ce",
     "https://deno.land/x/hono@v3.11.11/hono-base.ts": "4bc461d13072f48ebb3cf0be1651c114aa8c1bfbfffcd241cc9d18d23bdfedaf",
@@ -706,12 +708,15 @@
     "https://deno.land/x/hono@v3.11.11/router/trie-router/index.ts": "3eb75e7f71ba81801631b30de6b1f5cefb2c7239c03797e2b2cbab5085911b41",
     "https://deno.land/x/hono@v3.11.11/router/trie-router/node.ts": "1709750f3da7bf7fa3e8078caa1fc0bd5583b03c79f3a63f75480a2db25e62ad",
     "https://deno.land/x/hono@v3.11.11/router/trie-router/router.ts": "54ced78d35676302c8fcdda4204f7bdf5a7cc907fbf9967c75674b1e394f830d",
+    "https://deno.land/x/hono@v3.11.11/types.ts": "4cbc25194e89dcd78dd669e616150041f93eae6ad1825a601230bd922b568810",
     "https://deno.land/x/hono@v3.11.11/utils/body.ts": "7a16a6656331a96bcae57642f8d5e3912bd361cbbcc2c0d2157ecc3f218f7a92",
     "https://deno.land/x/hono@v3.11.11/utils/buffer.ts": "9066a973e64498cb262c7e932f47eed525a51677b17f90893862b7279dc0773e",
     "https://deno.land/x/hono@v3.11.11/utils/cookie.ts": "19920ba6756944aae1ad8585c3ddeaa9df479733f59d05359db096f7361e5e4b",
     "https://deno.land/x/hono@v3.11.11/utils/crypto.ts": "bda0e141bbe46d3a4a20f8fbcb6380d473b617123d9fdfa93e4499410b537acc",
     "https://deno.land/x/hono@v3.11.11/utils/html.ts": "b60def9561d2fbeb7254035632ea98b97ef852ec72baae2924fc126e96315696",
+    "https://deno.land/x/hono@v3.11.11/utils/http-status.ts": "e0c4343ea7717c314dc600131e16b636c29d61cfdaf9df93b267258d1729d1a0",
     "https://deno.land/x/hono@v3.11.11/utils/stream.ts": "1789dcc73c5b0ede28f83d7d34e47ae432c20e680907cb3275a9c9187f293983",
+    "https://deno.land/x/hono@v3.11.11/utils/types.ts": "41c4e7d8028f3dec43ff7c26e9498fed53d9ec7c3b47071d066b40f7b7a8ef58",
     "https://deno.land/x/hono@v3.11.11/utils/url.ts": "af0d600973ce11f9279da19a49e1264fe2f7a5658ccd97e30f7c3c2239dca052",
     "https://deno.land/x/hono@v3.11.11/validator/index.ts": "6c986e8b91dcf857ecc8164a506ae8eea8665792a4ff7215471df669c632ae7c",
     "https://deno.land/x/hono@v3.11.11/validator/validator.ts": "f6139f0b6eb2040bc9f7ed4148905d5affc898235952772b2d6585fdae8d4731",
