+++ src/config/types.ts
@@ -1,35 +1,35 @@
 /**
  * Configuration Types for Lens Engine
- * 
+ *
  * DENO MODULE ORGANIZATION PATTERN:
  * ================================
- * 
+ *
  * This file demonstrates the recommended structure for organizing types in a Deno
  * monorepo with multiple modules. Each module should follow this pattern:
- * 
+ *
  * module_name/
  * ├── src/           # Private implementation details
  * ├── types.ts       # Public type definitions (THIS FILE)
  * └── mod.ts         # Public API entry point
- * 
+ *
  * RATIONALE:
  * - types.ts sits at module level (same as mod.ts) to clearly indicate it's part
  *   of the public API, not internal implementation
  * - src/ directory contains all private implementation files
  * - mod.ts acts as the single entry point, re-exporting from both src/ and types.ts
  * - This creates clean module boundaries and makes imports predictable
- * 
+ *
  * USAGE PATTERN:
  * - Internal files import: import { SomeType } from "../types.ts"
  * - External consumers import: import { SomeType } from "./module_name/mod.ts"
  * - CLI and other modules get clean imports without deep paths into src/
- * 
+ *
  * This pattern scales well across a monorepo because:
  * 1. Each module is self-contained with clear public/private boundaries
  * 2. Type conflicts are avoided through proper module encapsulation
  * 3. Refactoring internal implementation doesn't break external consumers
  * 4. The module structure mirrors how external packages would be organized
- * 
+ *
  * EXAMPLE MODULE STRUCTURE:
  * src/
  * ├── config/          # This module
@@ -52,7 +52,7 @@
 
 /**
  * Core application configuration
- * 
+ *
  * Defines fundamental settings that apply across the entire Lens Engine,
  * including data storage locations, logging behavior, and server configuration.
  */
@@ -67,7 +67,7 @@ export interface CoreConfig {
 
 /**
  * Large Language Model configuration
- * 
+ *
  * Settings for LLM integration, including model selection and connection details.
  * Used by the processors module for content summarization and analysis.
  */
@@ -82,7 +82,7 @@ export interface LLMConfig {
 
 /**
  * LangChain framework configuration
- * 
+ *
  * Settings for LangChain integration, primarily for debugging and development.
  */
 export interface LangChainConfig {
@@ -92,7 +92,7 @@ export interface LangChainConfig {
 
 /**
  * LangSmith observability configuration
- * 
+ *
  * Settings for LangSmith integration to monitor and trace LLM operations.
  */
 export interface LangSmithConfig {
@@ -106,7 +106,7 @@ export interface LangSmithConfig {
 
 /**
  * Database configuration
- * 
+ *
  * Settings for data persistence, including local SQLite and optional vector database.
  */
 export interface DatabaseConfig {
@@ -118,7 +118,7 @@ export interface DatabaseConfig {
 
 /**
  * Complete application configuration
- * 
+ *
  * Root configuration interface that combines all module-specific configurations.
  * This is the primary interface used throughout the application.
  */
@@ -137,7 +137,7 @@ export interface AppConfig {
 
 /**
  * Configuration-related error
- * 
+ *
  * Custom error type for configuration loading and validation failures.
  * Provides better error handling and debugging for config-related issues.
  */
