--- src/processors/src/controller/content/mod.ts
@@ -1,15 +1,0 @@
-/**
- * Content Controllers Module
- *
- * Exports all content-specific processing controllers.
- */
-
-export {
-  type HtmlProcessingOptions,
-  processHtmlContent,
-  processHtmlFile,
-} from "./html.ts";
-
-// Future exports:
-// export { processMarkdownContent, processMarkdownFile } from "./markdown.ts";
-// export { processPdfContent, processPdfFile } from "./pdf.ts";
