--- src/processors/src/controller/processing/mod.ts
@@ -1,15 +1,0 @@
-/**
- * Processing Controllers Module
- *
- * Exports all processing orchestration controllers.
- */
-
-export {
-  type BatchProcessingOptions,
-  processHtmlBatch,
-  processMixedBatch,
-} from "./batch.ts";
-
-// Future exports:
-// export { processHtmlStream, processMixedStream } from "./streaming.ts";
-// export { processRealtime } from "./realtime.ts";
