+++ src/processors/src/html.ts
@@ -5,8 +5,8 @@
  * extraction → processing → output generation
  */
 
-import { extractFromHtml, type HtmlExtractResult } from "../../extract/mod.ts";
-import type { ProcessingOptions, ProcessingResult } from "../types.ts";
+import { extractFromHtml, type HtmlExtractResult } from "./extract.ts";
+import type { ProcessingOptions, ProcessingResult } from "./types.ts";
 import { summarizeContent, type SummaryOptions } from "@src/models/mod.ts";
 import { getConfig } from "@src/config/mod.ts";
 import { join } from "https://deno.land/std@0.224.0/path/mod.ts";
