+++ src/retrieval/mod.ts
@@ -1,11 +1,9 @@
 // retrieval/mod.ts
-export {
-  fetchAllContent,
-} from "./src/content_fetcher.ts";
+export { fetchAllContent } from "./src/content_fetcher.ts";
 
 export {
   type ContentFetcherOptions,
-  type FetchResult,
   type FetchOptions,
+  type FetchResult,
   type SaveOptions,
 } from "./types.ts";
