+++ src/retrieval/types.ts
@@ -1,21 +1,21 @@
 /**
  * Retrieval Module Types for Lens Engine
- * 
+ *
  * DENO MODULE ORGANIZATION PATTERN:
  * ================================
- * 
+ *
  * This file contains all public type definitions for the retrieval module.
  * Following the established pattern:
- * 
+ *
  * retrieval/
  * ├── src/           # Private implementation (content_fetcher.ts, etc.)
  * ├── types.ts       # Public type definitions (THIS FILE)
  * └── mod.ts         # Public API entry point
- * 
+ *
  * These types define the public interface for content fetching operations,
  * including configuration options, operation parameters, and result types.
  * Internal implementation details remain in src/ files.
- * 
+ *
  * USAGE:
  * - Internal files: import { FetchOptions } from "../types.ts"
  * - External consumers: import { FetchOptions } from "./retrieval/mod.ts"
@@ -24,7 +24,7 @@
 
 /**
  * Options for fetching content from a single URL
- * 
+ *
  * Configuration for individual fetch operations, including timeout and user agent settings.
  */
 export interface FetchOptions {
@@ -38,7 +38,7 @@ export interface FetchOptions {
 
 /**
  * Options for saving fetched content to disk
- * 
+ *
  * Configuration for file save operations, including path and overwrite behavior.
  */
 export interface SaveOptions {
@@ -50,7 +50,7 @@ export interface SaveOptions {
 
 /**
  * Main configuration options for the content fetcher
- * 
+ *
  * Primary interface used by the CLI and other consumers for batch content fetching.
  * This is the main public interface of the retrieval module.
  */
@@ -69,7 +69,7 @@ export interface ContentFetcherOptions {
 
 /**
  * Result of a single fetch operation
- * 
+ *
  * Contains the outcome of fetching content from a URL, including success status,
  * file path (if successful), or error information (if failed).
  */
